requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
plotly.express
scikit-learn
openpyxl
